ScoreComponent Enum to separate class
