+++ src/main/java/de/tum/mw/ftm/matsim/contrib/urban_ev/scoring/ScoreComponents.java
@@ -1,0 +1,16 @@
+package de.tum.mw.ftm.matsim.contrib.urban_ev.scoring;
+
+public enum ScoreComponents {
+
+        EMPTY_BATTERY,
+        WALKING_DISTANCE,
+        OPPORTUNITY_CHARGING, 
+        STATION_HOGGING, 
+        FAILED_CHARGING,
+        RANGE_ANXIETY,
+        HOME_CHARGING,
+        ENERGY_BALANCE,  
+        BATTERY_HEALTH,
+        CHARGING_EFFICIENCY
+    
+}
