+++ src/main/java/de/tum/mw/ftm/matsim/contrib/urban_ev/stats/ChargingScoreLogEntry.java
@@ -1,19 +1,19 @@
 package de.tum.mw.ftm.matsim.contrib.urban_ev.stats;
 
-import de.tum.mw.ftm.matsim.contrib.urban_ev.scoring.ChargingBehaviourScoring;
+import de.tum.mw.ftm.matsim.contrib.urban_ev.scoring.ScoreComponents;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.population.Person;
 
 public class ChargingScoreLogEntry {
     private final Id<Person> personId;
     private final double time; 
-    private final ChargingBehaviourScoring.ScoreComponents scoreComponent;
+    private final ScoreComponents scoreComponent;
     private final double value;
 
     public ChargingScoreLogEntry(
         Id<Person> personId,
         double time,
-        ChargingBehaviourScoring.ScoreComponents scoreComponent,
+        ScoreComponents scoreComponent,
         double value){
 
             this.personId = personId;
@@ -33,7 +33,7 @@ public class ChargingScoreLogEntry {
         return time;
     }
 
-    public ChargingBehaviourScoring.ScoreComponents getScoreComponent()
+    public ScoreComponents getScoreComponent()
     {
         return scoreComponent;
     }
